Updates for multi-version

+++ scripts/deploy.sh
@@ -24,19 +24,40 @@ if [ -z "${TRAVIS_TAG}" ]; then
 fi
 
 # Push the code to npm.
-npm publish --scope=@ibm
+BRANCH=`git symbolic-ref --short HEAD`
+if [ "${BRANCH}" = "develop" ]; then
+
+    # Publish with unstable tag. These are development builds.
+    echo "Pushing with tag unstable"
+    npm publish --scope=@ibm --tag=unstable
+
+else
+
+    # Publish with latest tag (default). These are release builds.
+    echo "Pushing with tag develop"
+    npm publish --scope=@ibm
+
+fi
+
+# If this is a tagged (release) build, don't go pushing to downstream projects.
+if [ -n "${TRAVIS_TAG}" ]; then
+    echo "Skipping downstream push; build is tagged"
+    exit 0
+fi
 
 # Push empty commits to downstream projects to trigger builds.
 REPO=`git config remote.origin.url`
 for PROJ in Concerto-Runtime; do
     cd ${DIR}
     THISREPO=$(echo ${REPO} | sed "s|/[^/]*$||")/${PROJ}.git
-    rm -rf temp
-    git clone ${THISREPO} temp
-    cd temp
-    git config user.email "noreply@ibm.com"
-    git config user.name "Blockchain WW Labs - Solutions"
-    git config push.default simple
-    git commit -m "Automated commit to trigger downstream build" --allow-empty
-    git push
+    for i in {1..5}; do
+        rm -rf temp
+        git clone -b ${BRANCH} ${THISREPO} temp
+        cd temp
+        git config user.email "noreply@ibm.com"
+        git config user.name "Blockchain WW Labs - Solutions"
+        git config push.default simple
+        git commit -m "Automated commit to trigger downstream build" --allow-empty
+        git push && break
+    done
 done
